add unit test for fluff case open range
