+++ server/src/tests/parser.spec.ts
@@ -1,8 +1,5 @@
-import { BailErrorStrategy, CommonTokenStream, PredictionMode } from "antlr4ng";
 import { DriverType } from "../backend/LpcConfig";
-import { LPCParser } from "../parser3/LPCParser";
-import { getLexer } from "./test-utils";
-import { LPCPreprocessingLexer } from "../parser3/LPCPreprocessingLexer";
+import { getParser } from "./test-utils";
 
 describe("Test", () => {
     beforeAll(() => {});
@@ -27,19 +24,3 @@ describe("Test", () => {
         expect(tree.children.length).toBeGreaterThan(0);
     });
 });
-
-function getParser(filename: string, driverType: DriverType) {
-    const lexer = getLexer(filename);
-    lexer.driverType = driverType;
-
-    const tokenStream = new CommonTokenStream(lexer);
-    tokenStream.fill();
-
-    const parser = new LPCParser(tokenStream);
-    parser.interpreter.predictionMode = PredictionMode.SLL;
-    parser.driverType = lexer.driverType;
-    parser.setTokenFactory(lexer.tokenFactory);
-    parser.buildParseTrees = true;
-    parser.errorHandler = new BailErrorStrategy();
-    return parser;
-}
