+++ server/src/tests/syntax.fluffos.spec.ts
@@ -1,0 +1,11 @@
+import { DriverType } from "../backend/LpcConfig";
+import { getParser } from "./test-utils";
+
+describe("FluffOS-specific Syntax Tests", () => {
+    it("should parse case statements with open-ended ranges", () => {
+        const parser = getParser("syntax-switch.fluffos.c", DriverType.FluffOS);
+        const tree = parser.program();
+
+        expect(tree).toBeDefined();
+    });
+});
